% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collection.R
\name{Collection}
\alias{Collection}
\alias{Categories}
\alias{Institutions}
\alias{Stats}
\alias{Collections}
\title{Retrieves Collection resources from the Symbiota2 server}
\usage{
Categories(url = NA, ID = NA, page = NA)

Institutions(url = NA, ID = NA, page = NA)

Stats(url = NA, ID = NA, page = NA)

Collections(url = NA, ID = NA, page = NA)
}
\arguments{
\item{url}{URL of the Symbiota2 portal connected to}

\item{ID}{ID value (usually \code{numeric}, but not always) 
used to refer to the specific resource to pull from the database}

\item{page}{\code{numeric} value referring to the page of Collection resources to pull. If neither an
ID or a page parameter is provided, function will pull the first page of resources (i.e. \code{page=1}`)}
}
\value{
If using \code{ID}, the specific Collection resource specified; 
if using page, the \code{page} specified of Collection resources
}
\description{
Functions that retrieve Collection resources from the server previously connected to.
Each function either retrieves an individual resource or a page of resources,
depending on the arguments provided.
}
\author{
Austin Koontz
}
