% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Different.R
\name{Descriptions}
\alias{Descriptions}
\title{requires multiple arguments for pulling specific resource}
\usage{
Descriptions(url = default.url, character.ID, characterState.ID, taxa.ID, page)
}
\description{
requires multiple arguments for pulling specific resource
}
