% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Occurrence.R
\name{Occurrences}
\alias{Occurrences}
\alias{AccessStats}
\alias{Determinations}
\alias{Duplicates}
\alias{EditLocks}
\alias{Edits}
\alias{FullText}
\alias{GuidDeterminations}
\alias{GuidOccurrences}
\alias{LookupChronostratigraphy}
\alias{LookupCounties}
\alias{LookupCountries}
\alias{LookupStateProvinces}
\alias{UploadMappings}
\alias{UploadParameters}
\alias{Verification}
\alias{Associations}
\alias{Comments}
\alias{DatasetLink}
\alias{Datasets}
\alias{Exchange}
\alias{Loans}
\title{Retrieves Occurrence resources from the Symbiota2 server}
\usage{
AccessStats(url = NA, ID = NA, page = NA)

Determinations(url = NA, ID = NA, page = NA)

Duplicates(url = NA, ID = NA, page = NA)

EditLocks(url = NA, ID = NA, page = NA)

Edits(url = NA, ID = NA, page = NA)

FullText(url = NA, ID = NA, page = NA)

GuidDeterminations(url = NA, ID = NA, page = NA)

GuidOccurrences(url = NA, ID = NA, page = NA)

LookupChronostratigraphy(url = NA, ID = NA, page = NA)

LookupCounties(url = NA, ID = NA, page = NA)

LookupCountries(url = NA, ID = NA, page = NA)

LookupStateProvinces(url = NA, ID = NA, page = NA)

UploadMappings(url = NA, ID = NA, page = NA)

UploadParameters(url = NA, ID = NA, page = NA)

Verification(url = NA, ID = NA, page = NA)

Associations(url = NA, ID = NA, page = NA)

Comments(url = NA, ID = NA, page = NA)

DatasetLink(url = NA, ID = NA, page = NA)

Datasets(url = NA, ID = NA, page = NA)

Exchange(url = NA, ID = NA, page = NA)

Loans(url = NA, ID = NA, page = NA)

Occurrences(url = NA, ID = NA, page = NA)
}
\arguments{
\item{url}{URL of the Symbiota2 portal connected to}

\item{ID}{ID value (usually \code{numeric}, but not always) used to refer to the specific resource to pull from the database}

\item{page}{\code{numeric} value referring to the page of Occurrence resources to pull. If neither an
ID or a page parameter is provided, function will pull the first page of resources (i.e. \code{page=1}`)}
}
\value{
If using \code{ID}, the specific Occurrence resource specified; If using page, the \code{page} specified
of Occurrence resources
}
\description{
Functions that retrieve Occurrence resources from the server previously connected to.
Each function either retrieves an individual resource or a page of resources,
depending on the arguments provided.
}
\author{
Austin Koontz
}
