% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Users.R
\name{Users}
\alias{Users}
\alias{UserRoles}
\title{Retrieves UserRoles resources from the Symbiota2 server}
\usage{
UserRoles(url = NA, ID = NA, permissions = NA)
}
\arguments{
\item{url}{URL of the Symbiota2 portal connected to}

\item{ID}{ID value (usually \code{numeric}, but not always) used to refer to the specific resource to pull from the database}

\item{permissions}{\code{numeric} value referring to permissions argument used to pull reource.}
}
\value{
UserRoles resource requested
}
\description{
Functions that retrieve UserRoles resources from the server previously connected to.
Each function either retrieves an individual resource or a page of resources,
depending on the arguments provided.
}
\author{
Austin Koontz
}
