% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Key.R
\name{Key}
\alias{Key}
\alias{CharacterHeading}
\alias{Characters}
\alias{CharacterStateImages}
\alias{CharacterStates}
\alias{DescriptionDeletions}
\title{Retrieves Key resources from the Symbiota2 server}
\usage{
CharacterHeading(id, page, url = NULL)

Characters(id, page, url = NULL)

CharacterStateImages(id, page, url = NULL)

CharacterStates(id, page, url = NULL)

DescriptionDeletions(id, page, url = NULL)
}
\arguments{
\item{id}{id value (usually \code{numeric}, but not always) 
used to refer to the specific resource to pull from the database}

\item{page}{\code{numeric} value referring to the page of resources to pull.
If neither an id or a page parameter is provided, 
function will pull the first page of resources (i.e. \code{page=1}`)}

\item{url}{URL string of the Symbiota2 portal to be connected to. 
A trailing \code{/} will be appended, if it is not given.}
}
\value{
If using \code{id}, the specific resource specified; 
if using page, the \code{page} specified of resources
}
\description{
Functions that retrieve Key resources from the server previously connected to.
Each function either retrieves an individual resource or a page of resources,
depending on the arguments provided.
}
\note{
To specify a default URL to refer to, see [SymbiotaR2_setup()]
}
\examples{
\dontrun{
# Pulling a Characters resource (id = 3), from a (nonexistent) dummy portal
object <- Characters(id = 3, url = "http://dummy-portal.com/api/")
}
}
\author{
Austin Koontz
}
