% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{Miscellaneous}
\alias{Miscellaneous}
\alias{Configurations}
\alias{Miscellaneous}
\alias{LookupLanguages}
\alias{Miscellaneous}
\alias{SchemaVersion}
\title{Retrieves Miscellaneous resources from the Symbiota2 server}
\usage{
Configurations(url = NA, ID = NA, page = NA)

LookupLanguages(url = NA, ID = NA, page = NA)

SchemaVersion(url = NA, ID = NA, page = NA)
}
\arguments{
\item{url}{URL of the Symbiota2 portal connected to}

\item{ID}{ID value (usually \code{numeric}, but not always) used to refer to the specific resource to pull from the database}

\item{page}{\code{numeric} value referring to the page of Miscellaneous resources to pull. If neither an
ID or a page parameter is provided, function will pull the first page of resources (i.e. \code{page=1}`)}
}
\value{
If using \code{ID}, the specific Miscellaneous resource specified; If using page, the \code{page} specified
of Miscellaneous resources
}
\description{
Functions that retrieve Miscellaneous resources from the server previously connected to.
Each function either retrieves an individual resource or a page of resources,
depending on the arguments provided.
}
\author{
Austin Koontz
}
