% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Occurrence.R
\name{Occurrences}
\alias{Occurrences}
\alias{AccessStats}
\alias{Determinations}
\alias{Duplicates}
\alias{EditLocks}
\alias{Edits}
\alias{FullText}
\alias{GuidDeterminations}
\alias{GuidOccurrences}
\alias{LookupChronostratigraphy}
\alias{LookupCounties}
\alias{LookupCountries}
\alias{LookupStateProvinces}
\alias{UploadMappings}
\alias{UploadParameters}
\alias{Verification}
\alias{Associations}
\alias{Comments}
\alias{DatasetLink}
\alias{Datasets}
\alias{Exchange}
\alias{Loans}
\title{Retrieves Occurrence resources from the Symbiota2 server}
\usage{
AccessStats(id, page, url = NULL)

Determinations(id, page, url = NULL)

Duplicates(id, page, url = NULL)

EditLocks(id, page, url = NULL)

Edits(id, page, url = NULL)

FullText(id, page, url = NULL)

GuidDeterminations(id, page, url = NULL)

GuidOccurrences(id, page, url = NULL)

LookupChronostratigraphy(id, page, url = NULL)

LookupCounties(id, page, url = NULL)

LookupCountries(id, page, url = NULL)

LookupStateProvinces(id, page, url = NULL)

UploadMappings(id, page, url = NULL)

UploadParameters(id, page, url = NULL)

Verification(id, page, url = NULL)

Associations(id, page, url = NULL)

Comments(id, page, url = NULL)

DatasetLink(id, page, url = NULL)

Datasets(id, page, url = NULL)

Exchange(id, page, url = NULL)

Loans(id, page, url = NULL)

Occurrences(id, page, url = NULL)
}
\arguments{
\item{id}{id value (usually \code{numeric}, but not always) 
used to refer to the specific resource to pull from the database}

\item{page}{\code{numeric} value referring to the page of resources to pull.
If neither an id or a page parameter is provided, 
function will pull the first page of resources (i.e. \code{page=1})}

\item{url}{URL string of the Symbiota2 portal to be connected to. 
A trailing \code{/} will be appended, if it is not given.}
}
\value{
If using \code{id}, the specific resource specified; 
if using page, the \code{page} specified of resources
}
\description{
Functions that retrieve Occurrence resources from the server previously connected to.
Each function either retrieves an individual resource or a page of resources,
depending on the arguments provided.
}
\note{
To specify a default URL to refer to, see \code{\link[SymbiotaR2:SymbiotaR2_setup]{SymbiotaR2_setup()}}
}
\examples{
\dontrun{
# Pulling a page of Occurrences, from a (nonexistent) dummy portal
entries <- Occurrence(page = 6, url = "http://dummy-portal.com/api/")
}
}
\author{
Austin Koontz
}
