% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Different.R, R/Traits.R
\name{Traits}
\alias{Traits}
\alias{Attributes}
\title{Retrieves Traits resources from the Symbiota2 server}
\usage{
Attributes(url = default.url, state.ID, occurrence.ID, page)

Traits(url = NA, ID = NA, page = NA)
}
\arguments{
\item{url}{URL of the Symbiota2 portal connected to}

\item{page}{\code{numeric} value referring to the page of Traits resources to pull. If neither an
ID or a page parameter is provided, function will pull the first page of resources (i.e. \code{page=1}`)}

\item{ID}{ID value (usually \code{numeric}, but not always) used to refer to the specific resource to pull from the database}
}
\value{
If using \code{ID}, the specific Traits resource specified; If using page, the \code{page} specified
of Traits resources
}
\description{
Functions that retrieve Traits resources from the server previously connected to.
Each function either retrieves an individual resource or a page of resources,
depending on the arguments provided.
}
\author{
Austin Koontz
}
